+++ HISTORY.rst
@@ -7,6 +7,7 @@ Historia zmian
 
 * wyświetlaj kolumnę z ilością wszystkich autorów w raporcie slotów - autor (#807)
 * wyświetlaj mniejsze czcionki w raporcie slotów - autor
+* raport slotów - ewaluacja (#809)
 
 202003.23
 ---------